pyproject.toml
[project]
name = "ai-trip-planner"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.116.1",
    "httpx>=0.28.1",
    "langchain>=0.3.27",
    "langchain-community>=0.3.27",
    "langchain-experimental>=0.3.4",
    "langchain-google-community[places]>=2.0.7",
    "langchain-groq>=0.3.7",
    "langchain-openai>=0.3.28",
    "langchain-tavily>=0.2.11",
    "langgraph>=0.6.4",
    "pandas>=2.3.0",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
    "requests>=2.32.4",
    "streamlit>=1.48.0",
    "uvicorn>=0.35.0",
]

[tool.uv.workspace]
members = [
    "AI_Trip_Planner",
]
